+++ test/node-tests/doc.test.js
@@ -38,6 +38,18 @@ function extractValue (spec) {
     words.shift()
     words[0] = 'DenseMatrix'
   }
+  // Collapse 'Complex Matrix' into 'ComplexMatrix', similarly for 'Array'
+  if (words[0] === 'Complex') {
+    if (words[1] === 'Matrix') {
+      words.shift()
+      words[0] = 'ComplexMatrix'
+    } else if (words[1] === 'Array') {
+      words.shift()
+      words[0] = 'ComplexArray'
+    } else if (words[1].startsWith('[')) {
+      words[0] = 'ComplexArray'
+    }
+  }
   const keywords = {
     number: 'Number(_)',
     BigNumber: 'math.bignumber(_)',
@@ -47,6 +59,8 @@ function extractValue (spec) {
     Array: '_',
     Matrix: 'math.matrix(_)',
     DenseMatrix: "math.matrix(_, 'dense')",
+    ComplexArray: 'math.complex(_)',
+    ComplexMatrix: 'math.complex(math.matrix(_))',
     string: '_',
     Node: 'math.parse(_)',
     value: 'math._',
@@ -92,15 +106,12 @@ const ignoreFunctions = new Set([
 ])
 
 const knownProblems = new Set([
-  'setUnion', 'unequal', 'equal', 'deepEqual', 'compareNatural', 'randomInt',
+  'setUnion', 'randomInt',
   'random', 'pickRandom', 'kldivergence',
   'parser', 'compile', 're', 'im',
   'subset', 'squeeze', 'rotationMatrix',
   'rotate', 'reshape', 'partitionSelect', 'matrixFromFunction',
-  'matrixFromColumns', 'getMatrixDataType', 'eigs', 'diff',
-  'nthRoots', 'nthRoot',
-  'mod', 'floor', 'fix', 'expm1', 'exp',
-  'ceil', 'cbrt', 'add', 'slu',
+  'matrixFromColumns', 'getMatrixDataType', 'eigs', 'diff', 'slu',
   'rationalize', 'qr', 'lusolve', 'lup', 'derivative',
   'symbolicEqual', 'schur', 'sylvester', 'freqz', 'round',
   'import', 'typed',
@@ -123,7 +134,12 @@ function maybeCheckExpectation (name, expected, expectedFrom, got, gotFrom) {
       }
     }
   } else {
-    checkExpectation(expected, got)
+    try {
+      checkExpectation(expected, got)
+    } catch (err) {
+      console.error(`DOC ERROR: '${gotFrom}' was supposed to '${expectedFrom}'`)
+      throw err
+    }
   }
 }
 
@@ -135,12 +151,15 @@ function checkExpectation (want, got) {
     return approxDeepEqual(got, want, 1e-9)
   }
   if (want instanceof math.Unit && got instanceof math.Unit) {
-    if (got.fixPrefix !== want.fixPrefix) {
+    if (got.fixPrefix !== want.fixPrefix ||
+      got.skipAutomaticSimplification !== want.skipAutomaticSimplification
+    ) {
       issueCount++
       if (debug) {
-        console.log('  Note: Ignoring different fixPrefix in Unit comparison')
+        console.log('  Note: Ignoring different flags in Unit comparison')
       }
       got.fixPrefix = want.fixPrefix
+      got.skipAutomaticSimplification = want.skipAutomaticSimplification
     }
     return approxDeepEqual(got, want, 1e-9)
   }
@@ -461,8 +480,13 @@ describe('Testing examples from (jsdoc) comments', function () {
                 if (accumulation) { accumulation += '\n' }
                 accumulation += parts[0]
               }
+              let discardAccumulation = true
               if (accumulation !== '' && expectation === undefined) {
                 expectationFrom = parts[1]
+                if (expectationFrom.endsWith('...')) {
+                  expectationFrom = expectationFrom.slice(0, -3)
+                  discardAccumulation = false
+                }
                 expectation = extractExpectation(expectationFrom)
                 parts[1] = ''
               }
@@ -476,7 +500,7 @@ describe('Testing examples from (jsdoc) comments', function () {
                 }
                 maybeCheckExpectation(
                   doc.name, expectation, expectationFrom, value, accumulation)
-                accumulation = ''
+                if (discardAccumulation) accumulation = ''
               }
               expectationFrom = parts[1]
               expectation = extractExpectation(expectationFrom, 'requireSignal')